requirements.txt
r2pipe==1.9.4
pefile==2023.2.7
python-magic==0.4.27
requests==2.32.3
python-dotenv==1.0.1
yara-python==4.5.1
pyfiglet==1.0.2